+++ remix/app/routes/post.tsx
@@ -1,0 +1,5 @@
+import { PostPage } from '~/components/Post/PostPage';
+
+export default function Post() {
+	return <PostPage />;
+}